Refactor serialization test cases so they can be reused
